put util functions into separate module
